# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LMS API Documentation",
        "description": "Apis to perform various actions on LMS resources",
        "contact": {
            "email": "support@email.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Test API Server"
        }
    ],
    "paths": {
...
        "/api/v1/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get all Jobs",
                "description": "Fetches a list of all available jobs",
                "operationId": "getJobs",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter jobs by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive",
                                "expired"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jobs Fetched Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Jobs Fetched Successfully"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "link": {
                                                        "type": "string"
                                                    },
                                                    "company_name": {
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "application_start_date": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "application_end_date": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "datetime"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "datetime"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Create a new job posting",
                "description": "Creates a new job posting with the provided details",
                "operationId": "storeJob",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "link",
                                    "company_name",
                                    "application_start_date",
                                    "application_end_date"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Senior Full Stack Developer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "We are looking for an experienced Full Stack Developer..."
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "https://company.com/careers/senior-dev"
                                    },
                                    "company_name": {
                                        "type": "string",
                                        "example": "Tech Corp Ltd"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "expired"
                                        ],
                                        "example": "active"
                                    },
                                    "application_start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-02"
                                    },
                                    "application_end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-04-02"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Job Created Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job Created Successfully"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "apiKey",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}